replaced text_wrapper to consider control sequences
